backend/requirements.txt
apsw==3.28.0.post1
apturl==0.5.2
asgiref==3.7.2
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
css-parser==1.0.4
cssselect==1.1.0
cssutils==1.0.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.4
django-allauth==0.55.0
django-cors-headers==4.2.0
django-cors-middleware==1.5.0
djangorestframework==3.14.0
dnspython==1.16.0
entrypoints==0.3
feedparser==5.2.1
galternatives==1.0.6
gunicorn==21.2.0
html2text==2020.1.16
html5-parser==0.4.9
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
ifaddr==0.1.6
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Markdown==3.1.1
mechanize==0.4.5
msgpack==0.6.2
netifaces==0.10.4
nose==1.3.7
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pychm==0.8.6
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
PyQt5==5.14.1
PyQtWebEngine==5.14.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
python3-openid==3.2.0
pytz==2019.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
regex==2019.8.19
reportlab==3.5.34
repoze.lru==0.7
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
Routes==2.4.1
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
soupsieve==1.9.5
sqlparse==0.4.4
systemd-python==234
typing-extensions==4.7.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
webencodings==0.5.1
WebOb==1.8.5
whitenoise==6.5.0
xkit==0.0.0
youtube-dl==2020.3.24
zeroconf==0.24.4
